moving jsxElementChildToText to dedicated file
