+++ editor/src/components/inspector/sections/component-section/component-section.tsx
@@ -109,7 +109,7 @@ import type { VariableData } from '../../../canvas/ui-jsx-canvas'
 import { array } from 'prop-types'
 import { useVariablesInScopeForSelectedElement } from './variables-in-scope-utils'
 import { DataPickerPopup } from './data-picker-popup'
-import { jsxElementChildToText } from '../../../canvas/ui-jsx-canvas-renderer/ui-jsx-canvas-element-renderer-utils'
+import { jsxElementChildToText } from '../../../canvas/ui-jsx-canvas-renderer/jsx-element-child-to-text'
 
 export const VariableFromScopeOptionTestId = (idx: string) => `variable-from-scope-${idx}`
 export const DataPickerPopupButtonTestId = `data-picker-popup-button-test-id`
